pyproject.toml
[project]
name = "branch-naming-hook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gitpython>=3.1.43",
]

[project.scripts]
branch-naming-hook = "branch_naming_hook:main"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.uv]
package = true
